requirements.txt
streamlit
plotly
spyder
python-keycloak
extra-streamlit-components
jwt